Refactor some tools (date, chart, tabs)
